Create dream
hi i have a great idea to add to open source app pianobooster http://pianobooster.sourceforge.net/ but i´m not a developer so i´m reaching out in hopes of finding the person who can make it happen. the idea is to make the notes being played on the midi instrument displayed like the note heads from midi files so that they scroll to the left as they are being played. the reason for this is that having the note heads you are playing scroll on the staff in real time might be a great way to learn how to read  because it is like learning to write pitches on the staff by simply playing music and and if you are learning how to write pitches you are subconsciously learning how to read music since the most difficult aspect of learning music is internalising pitches not rhythms. obviously i´m talking simply the note heads without stems. if you like the idea or know anyone who might be interested please let me know. thank you!